# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BlogO API",
    "description": "Auto-generated API documentation for the BlogO backend.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/"
    }
  ],
  "paths": {
...
    "/profile/": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userNiche": {
                    "example": "any"
                  },
                  "targetAudience": {
                    "example": "any"
                  },
                  "focusArea": {
                    "example": "any"
                  },
                  "productName": {
                    "example": "any"
                  },
                  "productDescription": {
                    "example": "any"
                  },
                  "productAudience": {
                    "example": "any"
                  },
                  "productSolution": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/generate-saas-profile": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/saas-profile": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}